sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atap-sdk"
version = "0.1.0"
description = "Python SDK for the ATAP (Agent Trust and Authority Protocol) platform"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "ATAP", email = "sdk@atap.dev" }]
keywords = ["atap", "agent", "trust", "delegation", "did", "verifiable-credentials"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "PyNaCl>=1.5.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "respx>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://atap.dev"
Documentation = "https://atap.dev/docs/sdks/python"
Repository = "https://github.com/8upio/atap"
Issues = "https://github.com/8upio/atap/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/atap"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.mypy]
python_version = "3.9"
strict = true

[tool.coverage.run]
source = ["atap"]

[tool.coverage.report]
fail_under = 80
